chore: refactor out common plugins
