+++ frontend/src/features/common/editor.ts
@@ -1,0 +1,34 @@
+import {
+  createBlockquotePlugin,
+  createBoldPlugin,
+  createCodeBlockPlugin,
+  createCodePlugin,
+  createHeadingPlugin,
+  createItalicPlugin,
+  createParagraphPlugin,
+  createPlateUI,
+  createPlugins,
+  createUnderlinePlugin,
+  PlatePlugin,
+} from '@udecode/plate'
+
+export const getCommonPlugins = (): PlatePlugin[] => {
+  const components = createPlateUI()
+  const plugins = createPlugins(
+    [
+      // elements
+      createParagraphPlugin(), // paragraph element
+      createBlockquotePlugin(), // blockquote element
+      createCodeBlockPlugin(), // code block element
+      createHeadingPlugin(), // heading elements
+
+      // marks
+      createBoldPlugin(), // bold mark
+      createItalicPlugin(), // italic mark
+      createUnderlinePlugin(), // underline mark
+      createCodePlugin(), // code mark
+    ],
+    { components },
+  )
+  return plugins
+}
