+++ frontend/src/pages/viewer/components/ReadonlyEditor.tsx
@@ -1,17 +1,7 @@
 import { FC } from 'react'
-import {
-  createBlockquotePlugin,
-  createBoldPlugin,
-  createCodeBlockPlugin,
-  createCodePlugin,
-  createHeadingPlugin,
-  createItalicPlugin,
-  createParagraphPlugin,
-  createPlateUI,
-  createPlugins,
-  createUnderlinePlugin,
-  Plate,
-} from '@udecode/plate'
+import { Plate } from '@udecode/plate'
+
+import { getCommonPlugins } from '~features/common/editor'
 
 interface ReadonlyEditorProps {
   value: string
@@ -20,23 +10,7 @@ interface ReadonlyEditorProps {
 export const ReadonlyEditor: FC<ReadonlyEditorProps> = ({ value }) => {
   const initialValue = JSON.parse(value)
 
-  const components = createPlateUI()
-  const plugins = createPlugins(
-    [
-      // elements
-      createParagraphPlugin(), // paragraph element
-      createBlockquotePlugin(), // blockquote element
-      createCodeBlockPlugin(), // code block element
-      createHeadingPlugin(), // heading elements
-
-      // marks
-      createBoldPlugin(), // bold mark
-      createItalicPlugin(), // italic mark
-      createUnderlinePlugin(), // underline mark
-      createCodePlugin(), // code mark
-    ],
-    { components },
-  )
+  const plugins = getCommonPlugins()
 
   return (
     <Plate
